coordinator/requirements.txt
flask>=3.0
pyserial>=3.5
librosa>=0.10
python-vlc>=3.0
